backend/requirements/dev.txt
-r base.txt

# Development
django-debug-toolbar>=4.4,<5.0
ipython>=8.27,<9.0
flake8>=7.1,<8.0
black>=24.8,<25.0
isort>=5.13,<6.0

# Testing
pytest>=8.3,<9.0
pytest-django>=4.9,<5.0
pytest-cov>=5.0,<6.0
factory-boy>=3.3,<4.0
faker>=28.0,<29.0
